+++ Assets/Scripts/Web/AppUser/AppUser.cs
@@ -5,13 +5,13 @@ using Sirenix.OdinInspector;
 public class AppUser
 {
     [ShowInInspector]
-    public int Id { get; set; }
+    public long Id { get; set; }
     [ShowInInspector]
     public String Code { get; set; }
-    public int WebSysUserId { get; set; }
+    public long WebSysUserId { get; set; }
     public String CSToken { get; set; }
     public int Options { get; set; } = 11;
-    public int ReferrerAppUserId { get; set; }
+    public long ReferrerAppUserId { get; set; }
     public int AppUserStatusId { get; set; }
 
 
@@ -19,7 +19,8 @@ public class AppUser
     {
     }
 
-    public AppUser(int id, String code, int webSysUserId, String phone, String csToken, int options, int referrerAppUserId, int appUserStatusId)
+    public AppUser(long id, String code, long webSysUserId, String phone, String csToken, int options,
+                   long referrerAppUserId, int appUserStatusId)
     {
         Id = id;
         Code = code;
