+++ Assets/Scripts/Web/AppUser/AppUserNamed.cs
@@ -5,10 +5,10 @@ using Sirenix.OdinInspector;
 public class AppUserNamed
 {
     [ShowInInspector]
-    public int Id { get; set; }
+    public long Id { get; set; }
     [ShowInInspector]
     public String Code { get; set; }
-    public int WebSysUserId { get; set; }
+    public long WebSysUserId { get; set; }
     [ShowInInspector]
     public String FirstName1 { get; set; }
     [ShowInInspector]
@@ -21,12 +21,12 @@ public class AppUserNamed
     public String LastNameMarried { get; set; }
     [ShowInInspector]
     public String Email { get; set; }
-    public int PhoneCountryId { get; set; } = -1;
+    public long PhoneCountryId { get; set; } = -1;
     [ShowInInspector]
     public String Phone { get; set; }
     public String CSToken { get; set; }
     public int Options { get; set; }
-    public int ReferrerAppUserId { get; set; }
+    public long ReferrerAppUserId { get; set; }
     public int AppUserStatusId { get; set; }
 
 
@@ -34,8 +34,8 @@ public class AppUserNamed
     {
     }
 
-    public AppUserNamed(int id, String code, int webSysUserId, String firstName1, String firstName2, String lastName1, String lastName2, String lastNameMarried,
-                        String email, int phoneCountryId, String phone, String cSToken, int options, int referrerAppUserId, int appUserStatusId)
+    public AppUserNamed(long id, String code, long webSysUserId, String firstName1, String firstName2, String lastName1, String lastName2, String lastNameMarried,
+                        String email, long phoneCountryId, String phone, String cSToken, int options, long referrerAppUserId, int appUserStatusId)
     {
         Id = id;
         Code = code;
