+++ Assets/Scripts/Web/Phone/PhoneCodeRequest.cs
@@ -2,7 +2,7 @@
 
 public class PhoneCodeRequest
 {
-    public int PhoneCountryId { get; set; }
+    public long PhoneCountryId { get; set; }
     public String PhoneNumber { get; set; }
     public int Code { get; set; }
 
@@ -11,7 +11,7 @@ public class PhoneCodeRequest
 
     }
 
-    public PhoneCodeRequest(int phoneCountryId, String phoneNumber, int code)
+    public PhoneCodeRequest(long phoneCountryId, String phoneNumber, int code)
     {
         PhoneCountryId = phoneCountryId;
         PhoneNumber = phoneNumber;
