+++ Assets/Scripts/Web/Referred/ReferredFull.cs
@@ -2,9 +2,9 @@ using System;
 
 public class ReferredFull
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public String Code { get; set; }
-    public int AppUserId { get; set; }
+    public long AppUserId { get; set; }
     public String FirstNames { get; set; }
     public String LastNames { get; set; }
     public String PhonePrefix { get; set; }
@@ -17,7 +17,7 @@ public class ReferredFull
     {
     }
 
-    public ReferredFull(int id, String code, int appUserId, String firstNames, String lastNames, String phonePrefix, String phone,
+    public ReferredFull(long id, String code, long appUserId, String firstNames, String lastNames, String phonePrefix, String phone,
                         String email, DateTime createDateTime, ReferrerFull referrer)
     {
         Id = id;
